code/requirements.txt
streamlit
pandas
numpy
plotly
langchain-openai
langchain-chroma
python-dotenv
statsmodels